Move methods from `grid-menu` and `grid-content` to `compile.service`
