+++ src/app/shared/compile/compile.service.ts
@@ -7,12 +7,14 @@ import {
     ViewContainerRef,
     ModuleWithProviders,
     Type,
-    Optional
+    Optional,
+    Inject
 } from '@angular/core';
 
 import { CommonModule } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from "@angular/router";
+import { MdDialog, MdDialogRef, MD_DIALOG_DATA } from '@angular/material';
 
 import { AuthService } from '../auth.service';
 import { WidgetService } from 'app/grid/widget.service';
@@ -83,7 +85,73 @@ export class CompileService  {
                 class TemplateComponent {
                     context: any
                     item = opts.item;
-                    constructor(private authService: AuthService, private widget: WidgetService) {}
+                    constructor(private authService: AuthService, private widget: WidgetService, public dialog: MdDialog) {}
+
+                    //WIDGET FUNCTIONS: TEMPORARY//
+
+                    url: string;
+                    engines = ["Google", "Bing", "DuckDuckGo"];
+                    clocks = ["AnalogWhite", "AnalogGreen", "DigitalBlue"];
+
+                    getURL(url: string): string {
+                        return "http://api.screenshotlayer.com/api/capture?access_key=a2f073b50b57b8c177482fa83b336efc&url=" + url;
+                    }
+
+                    refresh() {
+                        location.reload();
+                    }
+
+                    searchFor(value: string, item: any) {
+                        if (value !== "" && value !== undefined && value !== null)
+                            if (item.setting !== "DuckDuckGo")
+                                window.location.href = 'https://' + item.setting + '.com/search?q=' + value;
+                        if (item.setting === "DuckDuckGo")
+                            window.location.href = 'https://' + item.setting + '.com/?q=' + value;
+                    }
+
+                    changeURL() {
+                        this.openDialog();
+                        alert("Your Current URL Setting: " + this.item.setting);
+                        var txt;
+                        var url = prompt("Please enter the new website's url:");
+                        if (url == null || url == "") {
+                            alert("No changes has been made.");
+                            return;
+                        } else {
+                            txt = url;
+                            this.item.setting = url;
+                        }
+                    }
+
+
+                    openDialog(): void {
+                        let dialogRef = this.dialog.open(ChangeURLDialog, {
+                            width: '250px',
+                            data: { url: this.url }
+                        });
+                        dialogRef.componentInstance.dialogRef = dialogRef;
+
+                        dialogRef.afterClosed().subscribe(result => {
+                            console.log('The dialog was closed');
+                            this.url = result;
+                        });
+                    }
+                }
+
+                @Component({
+                    selector: 'changeURL-dialog',
+                    templateUrl: '../../customize/grid-menu/changeURLDialog.html',
+                })
+                class ChangeURLDialog {
+
+                    constructor(
+                        public dialogRef: MdDialogRef<ChangeURLDialog>,
+                        @Inject(MD_DIALOG_DATA) public data: any) { }
+
+                    onNoClick(): void {
+                        this.dialogRef.close();
+                    }
+
                 }
 
                 let module : NgModule = {};
@@ -96,6 +164,8 @@ export class CompileService  {
                 module.imports.push( CommonModule );
                 module.imports.push( BrowserModule ); 
                 module.imports.push( RouterModule ); 
+                module.providers = module.providers || [];
+                module.providers.push( MdDialog );
                 if (opts.imports !== undefined) {
                     module.imports = module.imports.concat(opts.imports)
                 }
